test-headerbar: create a GtkShortcutsWindow manually
Quick test to see if it's possible to create a GtkShortcutsWindow with
code (not with a *.ui file). And the answer is yes, it's possible, it
works. Note that all the GtkShortcuts* APIs don't have any public
function, the documentation says that it's recommended to use a *.ui
file :-(

There is the GtkShortcutsShortcut:action-name property, but the API
doesn't say whether it takes only the first accel or all accels. In
Devhelp I want to take only the first accel if there are several.

Another advantage with Amtk is that all the information is centralized,
the GtkShortcutsShortcut:title will be defined in the AmtkActionInfo, so
when modifying one piece of info we see all the other related info for
the same action. So when renaming the label, we think about renaming the
title for the GtkShortcutsShortcut too, since the two strings will be at
the same place.