build: adapt for AX_CODE_COVERAGE API break
The usage of AX_CODE_COVERAGE has changed since autoconf-archive version
2019.01.06.

Unfortunately the ax_code_coverage.m4 from version 2019.01.06 has a bug:
```
*** Recursive variable 'GITIGNOREFILES' references itself (eventually).  Stop.
```
which is fixed in autoconf-archive git master, but not yet in a stable
version of autoconf-archive.
--> So have a local copy of ax_code_coverage.m4 with the bug fixes, in the m4/
directory.

And add aminclude_static.am to the MAINTAINERCLEANFILES, so that git.mk
will add that file to the .gitignore.

Fixes https://gitlab.gnome.org/GNOME/amtk/issues/1