build: Support Visual Studio builds with NMake
This adds a set of NMake Makefiles that can be used to build amtk with its
tests and sample programs, and optionally with the introspection files.

A README.txt file, along with an information page that is displayed when
running 'nmake /f Makefile.vc' is invoked without the CFG parameter, which
reveals the options that must be passed to 'nmake /f Makefile.vc' to carry
out the build, as well as optional options that can be used as well.

This will also produce a working pkg-config file if Python was used during
the build/install process, and a 'test' target is also supported to run the
test programs in testsuite/.